requirements.txt
Django==1.11.2
djangorestframework==3.7.3
MySQL-python==1.2.5
Pillow==4.3.0
django-crispy-forms
future
django-reversion==2.0.12
django-formtools==2.1.0
django_import_export==0.6.1
xlwt==1.3.0
XlsxWriter==1.0.2
django_guardian==1.4.9
coreapi==2.3.3
httplib2==0.10.3
django-filter==1.1.0
djangorestframework-jwt==1.11.0




